OS X 10.7 w/ Xcode 4.3 does not include libpng any longer, so I added it back into the G3D build.  This will be replaced with FreeImage shortly anyway.
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1258 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/source/fileutils.cpp
@@ -148,7 +148,7 @@ std::string readWholeFile(const std::string& filename) {
                 throw std::string("\"") + internalFile + "\" inside \"" + zipfile + "\" could not be opened.";
             } else {
                 const int64 bytesRead = zip_fread( zf, buffer, (info.size));
-                debugAssertM(bytesRead == info.size,
+                debugAssertM((int64)bytesRead == info.size,
                              internalFile + " was corrupt because it unzipped to the wrong size.");
                 (void)bytesRead;
                 zip_fclose( zf );
